+++ packages/core/legos/CHANGELOG.md
@@ -5,9 +5,9 @@
 ### Patch Changes
 
 - Updated dependencies [e39eb6e6c]
-- Updated dependencies
+- Updated dependencies [5258e1166]
   - @chainlink/intrinio-test-adapter@1.0.1
-  - @chainlink/twosigma-adapter@1.0.0
+  - @chainlink/coinranking-adapter@2.0.0
 
 ## 1.3.63